@@ -14,13 +14,21 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
+
+  google_sign_in: ^5.0.3
+  firebase_auth: ^2.0.0
+  cloud_firestore: ^2.2.0
+  firebase_core: ^1.2.0
+  firebase_storage: ^9.0.0
+  toast: ^0.1.5
   cupertino_icons: ^0.1.3
-  flutter_blue: ^0.7.2
+  flutter_blue: ^0.8.0
+  flame: 1.0.0-rc9
 
 
   # The following adds the Cupertino Icons font to your application.
@@ -41,8 +49,16 @@ flutter:
   # the material Icons class.
   uses-material-design: true
 
-assets:
-    - assets/images/temp.png;
+  assets:
+    - bluewhite.png
+    - Dino.png
+    - Fish.jpg
+    - Pump.jpg
+    - Bluewhite_text.png
+    - Dino_text.png
+    - Fish_text.png
+    - Pump_text.png
+    - sprite.png
 
   # To add assets to your application, add an assets section, like this:
   # assets:
