@@ -4,6 +4,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_blue/flutter_blue.dart';
 import 'package:move/data.dart';
 import 'package:move/value.dart';
+
+import 'front/login.dart';
+
 String gesture = "";
 // ignore: non_constant_identifier_names
 int gesture_num = 0;
@@ -18,21 +21,14 @@ class MyHomePage extends StatefulWidget {
 
 
 class _MyHomePageState extends State<MyHomePage> {
-  Gesture _gesturedata;
-  Gesture get gesturedata{
-    if (_gesturedata == null) {
-      _gesturedata = Gesture(0); // Instantiate the object if its null.
-    }
-    return _gesturedata;
-  }
-
+  final Move move = Move(gesture_num);
   final FlutterBlue flutterBlue = FlutterBlue.instance;
   // ignore: deprecated_member_use
-  final List<BluetoothDevice> devicesList = new List<BluetoothDevice>();
+  final List<BluetoothDevice> devicesList = [];
   final Map<Guid, List<int>> readValues = new Map<Guid, List<int>>();
   final textController = TextEditingController();
-  BluetoothDevice connectedDevice;
-  List<BluetoothService> bluetoothServices;
+  BluetoothDevice? connectedDevice;
+  List<BluetoothService>? bluetoothServices;
 
   _showDeviceTolist(final BluetoothDevice device) {
     if (!devicesList.contains(device)) {
@@ -62,7 +58,7 @@ class _MyHomePageState extends State<MyHomePage> {
 
   ListView _buildListViewOfDevices() {
     // ignore: deprecated_member_use
-    List<Container> containers = new List<Container>();
+    List<Container> containers = [];
     for (BluetoothDevice device in devicesList) {
       containers.add(
         Container(
@@ -89,7 +85,7 @@ class _MyHomePageState extends State<MyHomePage> {
                   try {
                     await device.connect();
                   } catch (e) {
-                    if (e.code != 'already_connected') {
+                    if (e != 'already_connected') {
                       throw e;
                     }
                   } finally {
@@ -117,7 +113,7 @@ class _MyHomePageState extends State<MyHomePage> {
   List<ButtonTheme> _buildReadWriteNotifyButton(
       BluetoothCharacteristic characteristic) {
     // ignore: deprecated_member_use
-    List<ButtonTheme> buttons = new List<ButtonTheme>();
+    List<ButtonTheme> buttons = [];
     if (characteristic.properties.notify) {
       buttons.add(
         ButtonTheme(
@@ -166,9 +162,7 @@ class _MyHomePageState extends State<MyHomePage> {
         readValues[characteristic.uuid] = value;
         gesture = value.toString();
         gesture_num = int.parse(gesture[1]);
-        gesturedata.gesturedata = 0;
-        gesturedata.gesturedata = gesture_num;
-        switch(gesturedata.gesturedata){
+        switch(gesture_num){
           case 1: gesture_name = "LEFT"; break;
           case 2: gesture_name = "RIGHT"; break;
           case 3: gesture_name = "UP"; break;
@@ -180,12 +174,13 @@ class _MyHomePageState extends State<MyHomePage> {
     await characteristic.read();
     sub.cancel();
   }
+
   ListView _buildConnectDeviceView() {
     // ignore: deprecated_member_use
-    List<Container> containers = new List<Container>();
-    for (BluetoothService service in bluetoothServices) {
+    List<Container> containers = [];
+    for (BluetoothService service in bluetoothServices!) {
       // ignore: deprecated_member_use
-      List<Widget> characteristicsWidget = new List<Widget>();
+      List<Widget> characteristicsWidget = [];
 
       for (BluetoothCharacteristic characteristic in service.characteristics) {
         characteristicsWidget.add(
@@ -217,20 +212,20 @@ class _MyHomePageState extends State<MyHomePage> {
       children: <Widget>[
         Center(child:containers[2]),
         Container(
-        child:Column(
-          children: [
-            SizedBox(height: 30,),
-            Center(
-                child:Column(
-                  children: [
-                    Text("값:" + gesture_num.toString(),style: TextStyle(fontSize: 20,fontWeight: FontWeight.bold),),
-                    SizedBox(height: 30,),
-                    Text(gesture_name,style: TextStyle(fontSize: 20,fontWeight: FontWeight.bold),),
-                  ],
-                )
-              ),
-          ],
-        )
+            child:Column(
+              children: [
+                SizedBox(height: 30,),
+                Center(
+                    child:Column(
+                      children: [
+                        Text("값:" + gesture_num.toString(),style: TextStyle(fontSize: 20,fontWeight: FontWeight.bold),),
+                        SizedBox(height: 30,),
+                        Text(gesture_name,style: TextStyle(fontSize: 20,fontWeight: FontWeight.bold),),
+                      ],
+                    )
+                ),
+              ],
+            )
         ),
       ],
     );
@@ -245,30 +240,39 @@ class _MyHomePageState extends State<MyHomePage> {
 
   @override
   Widget build(BuildContext context) => Scaffold(
-        appBar: AppBar(
-          title: Center(child: Text("MOVE!")),
-            leading: IconButton(
-              icon: Icon(Icons.people_alt,color: Colors.white,),
-              onPressed: () => {
-                print(gesturedata.gesturedata),
-              },
-            ),
-            actions: <Widget>[
-              new IconButton(
-                icon: new Icon(Icons.photo_album),
-                tooltip: 'Hi!',
-                onPressed: () => {
-                  print(gesture_num),
-                Navigator.push(
+    appBar: AppBar(
+      title: Center(child: Text("MOVE!")),
+      leading: IconButton(
+        icon: Icon(Icons.people_alt,color: Colors.white,),
+        onPressed: () => {
+
+        },
+      ),
+      actions: <Widget>[
+        new IconButton(
+          icon: new Icon(Icons.photo_album),
+          tooltip: 'Hi!',
+          onPressed: () async {
+            print(gesture_name);
+            final move = Move(gesture_num);
+            await Navigator.push(
                 context,
-                MaterialPageRoute(builder: (context) => CounterPage(gesturedata)),
-                )
-                },
-              ),
-            ],
+                MaterialPageRoute(builder: (context) => CounterPage(bluetoothServices: bluetoothServices))
+                // MaterialPageRoute(builder: (context) => CounterPage(move: move))
+            );
+          },
         ),
-        body: _buildView(),
-      );
+        IconButton(
+                icon: new Icon(Icons.design_services),
+                onPressed: () => {
+                      print(gesture_num),
+                      Navigator.push(
+                        context,
+                        MaterialPageRoute(builder: (context) => Login()),
+                      )
+                    })
+          ],
+    ),
+    body: _buildView(),
+  );
 }
-
-
