@@ -1,10 +1,14 @@
-
-
+import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
+import '/front/login.dart';
 
 import '/home_page.dart';
 
-void main() => runApp(MyApp());
+void main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+  await Firebase.initializeApp();
+  runApp(MyApp());
+}
 
 class MyApp extends StatelessWidget {
   @override
